chore: refactor paperinputpicker to include label
